refactor: move common utilities into indexer-common crate
