--- native/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "native"
-version = "0.1.0"
-authors = ["Zac Burns <That3Percent@gmail.com>"]
-license = "MIT"
-edition = "2018"
-exclude = ["artifacts.json", "index.node"]
-
-# [lib]
-# name = "native"
-# crate-type = ["cdylib"]
-
-# [build-dependencies]
-# neon-build = "0.9.1"
-
-[dependencies]
-# neon = "0.9"
-# neon-utils = { git = "https://github.com/edgeandnode/neon-utils", rev = "2507d4f" }
-secp256k1 = { version = "0.20", features = ["recovery"] }
-never = "0.1"
-keccak-hash = "0.5.1"
-lazy_static = "1.4"
-arc-swap = "1.2"
-eip-712-derive = { git = "https://github.com/graphprotocol/eip-712-derive" }
-hex = "0.4.2"
-primitive-types = "0.8"
