+++ service/src/common/mod.rs
@@ -2,9 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod address;
-pub mod allocation;
 pub mod database;
 pub mod indexer_error;
 pub mod indexer_management;
-pub mod network_subgraph;
-pub mod types;
